cleanup: clippy --fix machine-applicable lints in storage crates
Mechanical, semantics-preserving fixes (Error::other, div_ceil, redundant
closures, needless casts, etc.) across consus-hdf5/consus-io/consus-zarr — 28
fixes in 13 files. Pre-existing debt unrelated to the S3 work but blocking the
clippy -D warnings gate. Verified: 1166 default-feature tests pass + s3-moirai green.

+++ crates/consus-zarr/src/codec/mod.rs
@@ -180,13 +180,13 @@ impl CodecPipeline {
                     .and_then(|c| c.compress(data, CompressionLevel(level)))
             }
             "zstd" => {
-                let level = codec.zstd_level().unwrap_or(3) as i32;
+                let level = codec.zstd_level().unwrap_or(3);
                 registry
                     .get_by_name(&codec.name)
                     .and_then(|c| c.compress(data, CompressionLevel(level)))
             }
             "lz4" => {
-                let level = codec.lz4_level().unwrap_or(0) as i32;
+                let level = codec.lz4_level().unwrap_or(0);
                 registry
                     .get_by_name(&codec.name)
                     .and_then(|c| c.compress(data, CompressionLevel(level)))
